fix(neon-core): Fix ContractParam.Integer transformation (#347)
C# BigInteger uses two's complement which wasn't reflected by the `parseInt` of javascript. This solution brings in Igor's `csbiginteger` package which is a javascript implementation of the c# BigInteger class. With this, we should be able to convert correctly. It's a good place to start testing the package out too.

Fix #345