fix(sc): StackItem serialization of zero values (#394)
When the VM creates the OpCodes for StackItems of null, false or zero values, the OpCode returned is 0000. This means to read off zero bytes. Unfortunately, our implementation of StringStream does not check if it reads zero and throws an error if attempting to read from the end of the stream.

This changes the implementation of StackItem to first check if we are reading zero bytes. If we are, use a default value as indicated by the StackItemType.