feat(tx): Use HexString class in tx (#496)
This replaces the hexstring fields in the tx folder to use HexString class in u. This should allow more checking and easier development on our side. It also gives us easy 0x stripping, hex checks and removes endian knowledge.

One downside is that it is slightly more messy to write tests. Just like the Fixed8 construct, the string is hidden and we have to keep calling equal or toBigEndian to test the inner values.