fix(sc): Fix emitAppCall with only scriptHash (#506)
fixes #502

Originally, the call to emitAppCall without any arguments will produce a script that will push a 00 at the start of the script. This can be seen as an error if the contract did not consume the parameter, leaving it in the stack.

By adding a check for undefined, we can prevent the 00 from appearing when no arguments are provided. This breaks existing behaviour for all scripts that rely on no arguments filling in an empty parameter. For example, standard NEP5 contracts will consume 2 variables off the stack. This fix will cause the old code to emit only a single variable and result in a fault when running the script.

This is fixed by explicitly providing an empty array for args (which is what should happen anyway) which will prompt the ScriptBuilder to inject the appropriate variable onto the stack.