feat(u): Fixed8.fromRawNumber (#541)
NEO RPCs sometimes return GAS values as raw numbers like "1000000000" which are sometimes too big to fit within the Fixed8 constrain (Not technical limitations but we would like to make sure every Fixed8 is within the max and min). To resolve this, the `fromRawNumber` method is a way for us to construct Fixed8 from such large raw values.