feat(sc): OpToken (#577)
OpToken is a tokenized form of a script. This is generated from parsing a VM script and breaking it down into tokens and its params if any. The core motivation here is to have a better system that the current ScriptParser in order to correctly breakdown the multisig account script.