feat(sc): A SmartContract base class (#612)
A SmartContract framework for simple contract script construction. 

Before, the old framework revolves around the interface `ScriptIntent` (which is very similar to the new `ContractCall` interface). This was designed to be a simple object that SDK users can create to denote a smart contract method call. However, it was not really fully supported by the old `neon-nep5` package, which was coupled tightly with the ScriptBuilder. The `args` field was also `unknown`, leading to multiple parsing issues as there was insufficient information to verify and serialize the data.

This PR reworks `ScriptIntent` and the `neon-nep5` package for better support:

`ScriptIntent` is now superceded by `ContractCall`. With NEO3 standardizing the entry point, we can now support the (scriptHash, operation, args) inputs as a first class citizen for all contracts. Args is now typed as ContractParams which provided stable typing for us to reliably execute serialization.

The `BaseContract` class provides a base layer for constructing a simple SmartContract representation that can be used to safely construct `ContractCall` objects with their parameters parsed correctly.

The `Nep5Contract` class is an example of how the `BaseContract` class is extended and provides a clean API interface. I will further extend it to implement the NEO and GAS contracts.

Translation of a `ContractCall` into a hexstring is trivial with ScriptBuilder as all the parameters are now well defined. The `ContractCall` object can also be easily used with `invokeFunction` RPC call.

I intend to deprecate:

- ContractParam.byteArray(value, format,...args) -> ContractParam.byteArray(value). I think moving the parameter translation upwards would reduce the burden on this class. For address conversions, there is `ContractParam.hash160`. For Fixed8 conversion, we should rely on BigInteger class. I plan to remove the 2 extra arguments.

- ScriptBuilder.emitAppCall. This can be superceded by a new method to accept `ContractCall`. For now, I plan to only mark it as deprecated and visit this again after more implementation.