feat(u): BigInteger.fromDecimal (#616)
In order to transit from Fixed8, the SDK will need to be able to parse decimal numbers. The new `BigInteger` class does not support decimals correctly. This introduces methods on BigInteger to work with decimals numbers. It effectively transforms the decimal numbers into a BigInteger by shifting the decimal place.