feat(api): NetworkFacade (#656)
NetworkFacade is the new high level API interface replacing the old methods of doInvoke,sendAsset and claimGas. It is intended to be a quick way to perform common wallet operations as well as a note to developers on how to use the SDK. For now, the only implementation is the replacement for sendAsset.