fix(sc): Fix ContractPermissions toJson dropping 0x (#758)
ContractPermissionDescriptor uses the UIint classes which maintains the 0x prefix when exporting to JSON. It is possible for the C# code to perform a safe parse using TryParse but currently, it uses a string length check to determine the data type. Thus, a non-prefixed string  fails the length check and errors out. This fixes the export to JSON to maintain the prefix for contract hashes.

Fix #756