feat(uri): neon-uri (#772)
neon-uri is a rename and upgrade of nep-9. Nep9 style uris are very specific to neo2 and are not too relevant in N3 context. neon-uri retains the core support of token transfers while also adding an additional voting uri support. It draws from ethereum's proposal of adding prefixes to indicate the core intent of the uri.

- Rename neon-nep9 to neon-uri.
- Update parse to support voting.
- Add uri creation methods.