feat(rpc): BatchQuery (#812)
BatchQuery allows for an array of jsonrpc calls to be sent over a single rpc call instead of trying to join them together. The batching mechanism is designed with the type system so that the return values are correctly typed. However, this means that the system is forced to throw when any error occurs.
If your request has a danger of failing, use the `sendQueryList` method directly as it does not handle the typing or response unwrapping.

- Add BatchQuery
- Add RpcDispatcher.executeAll and sendQueryList methods